+++ NangPaGo-api/NangPaGo-app/src/main/java/com/mars/app/auth/handler/OAuth2SuccessHandler.java
@@ -3,7 +3,7 @@ package com.mars.app.auth.handler;
 import static com.mars.common.exception.NPGExceptionType.NOT_FOUND_USER;
 
 import com.mars.app.auth.vo.OAuth2UserImpl;
-import com.mars.common.util.JwtUtil;
+import com.mars.common.util.web.JwtUtil;
 import com.mars.app.domain.auth.service.OAuth2ProviderTokenService;
 import com.mars.app.domain.auth.service.TokenService;
 import com.mars.common.model.user.User;
