+++ NangPaGo-api/NangPaGo-app/src/test/java/com/mars/app/provider/TestJwtProvider.java
@@ -1,6 +1,6 @@
 package com.mars.app.provider;
 
-import com.mars.common.util.JwtUtil;
+import com.mars.common.util.web.JwtUtil;
 import org.springframework.stereotype.Component;
 
 @Component
